Move `ty::Clause` to its own little module (cute)
